refactor: move Color file and change imports accordingly
