+++ app/src/main/kotlin/view/Color.kt
@@ -1,4 +1,4 @@
-package view.theme
+package view
 
 import androidx.compose.material3.darkColorScheme
 import androidx.compose.material3.lightColorScheme
